refactor : AppDelegate classを別のファイルに分ける
